[Refactoring] split API views with Flask blueprint
